update unit test code
